requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

torch==2.2.2+cpu
torchvision==0.17.2+cpu

numpy<2.0
opencv-python-headless

fastapi
uvicorn
python-multipart
streamlit
pillow
tqdm
kagglehub
scikit-learn
jinja2
